+++ app/profile/activate/loading.js
@@ -1,0 +1,10 @@
+import CommonLayout from '../../CommonLayout'
+import LoadingSpinner from '../../../components/LoadingSpinner'
+
+export default function Loading() {
+  return (
+    <CommonLayout banner={null}>
+      <LoadingSpinner />
+    </CommonLayout>
+  )
+}
